+++ db/src/main/java/dao/UserDao.java
@@ -5,7 +5,7 @@ import role.Role;
 import user.User;
 
 public interface UserDao {
-    User getUserById( long userId ) throws UserNotFoundException;
-    boolean deleteUserById( long userId ) throws UserNotFoundException;
+    User getUserById( int userId ) throws UserNotFoundException;
+    boolean deleteUserById( int userId ) throws UserNotFoundException;
     boolean registerUser( String username, String password, Role role );
 }
